requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
langchain>=0.3.0
langchain-core>=0.3.0
langchain-google-vertexai>=2.0.0
google-cloud-aiplatform>=1.60.0
python-dotenv>=1.0.0
pydantic>=2.0.0
